+++ mlflow/statsmodels.py
@@ -28,7 +28,7 @@ from mlflow.utils.environment import _mlflow_conda_env
 from mlflow.utils.model_utils import _get_flavor_configuration
 from mlflow.exceptions import MlflowException
 from mlflow.utils.annotations import experimental
-from mlflow.utils.autologging import (
+from mlflow.utils.autologging_utils import (
     try_mlflow_log,
     log_fn_args_as_params,
     autologging_integration,
