+++ tests/autologging/test_autologging_safety_integration.py
@@ -8,7 +8,7 @@ from itertools import permutations
 import mlflow
 from mlflow.tracking import MlflowClient
 from mlflow.utils import gorilla
-from mlflow.utils.autologging import (
+from mlflow.utils.autologging_utils import (
     safe_patch,
     get_autologging_config,
     autologging_is_disabled,
