+++ tests/lightgbm/test_lightgbm_autolog.py
@@ -12,7 +12,7 @@ import mlflow
 import mlflow.lightgbm
 from mlflow.models import Model
 from mlflow.models.utils import _read_example
-from mlflow.utils.autologging import BatchMetricsLogger
+from mlflow.utils.autologging_utils import BatchMetricsLogger
 from unittest.mock import patch
 
 mpl.use("Agg")
@@ -248,7 +248,7 @@ def test_lgb_autolog_batch_metrics_logger_logs_expected_metrics(bst_params, trai
     original = BatchMetricsLogger.record_metrics
 
     with patch(
-        "mlflow.utils.autologging.BatchMetricsLogger.record_metrics", autospec=True
+        "mlflow.utils.autologging_utils.BatchMetricsLogger.record_metrics", autospec=True
     ) as record_metrics_mock:
 
         def record_metrics_side_effect(self, metrics, step=None):
