+++ tests/tensorflow_autolog/test_tensorflow2_autolog.py
@@ -12,7 +12,7 @@ from tensorflow.keras import layers
 import mlflow
 import mlflow.tensorflow
 import mlflow.keras
-from mlflow.utils.autologging import BatchMetricsLogger
+from mlflow.utils.autologging_utils import BatchMetricsLogger
 from unittest.mock import patch
 
 import os
@@ -374,7 +374,7 @@ def test_tf_keras_autolog_batch_metrics_logger_logs_expected_metrics(
     original = BatchMetricsLogger.record_metrics
 
     with patch(
-        "mlflow.utils.autologging.BatchMetricsLogger.record_metrics", autospec=True
+        "mlflow.utils.autologging_utils.BatchMetricsLogger.record_metrics", autospec=True
     ) as record_metrics_mock:
 
         def record_metrics_side_effect(self, metrics, step=None):
